+++ spec/parser/escapes_spec.rb
@@ -38,6 +38,7 @@ RSpec.describe('EscapeSequence parsing') do
   include_examples 'parse', /\?/,            0 => [char:  '?',     codepoint:  63      ]
   include_examples 'parse', /\101/,          0 => [char:  'A',     codepoint:  65      ]
   include_examples 'parse', /\x42/,          0 => [char:  'B',     codepoint:  66      ]
+  include_examples 'parse', /\xA/,           0 => [char:  "\n",    codepoint:  10      ]
   include_examples 'parse', /\u0043/,        0 => [char:  'C',     codepoint:  67      ]
   include_examples 'parse', /\u{44 45}/,     0 => [chars: %w[D E], codepoints: [68, 69]]